requirements.txt
# CORE DEEP LEARNING
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# DATA HANDLING
datasets>=2.18.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# VISUALIZATION & UI
matplotlib>=3.7.0
streamlit>=1.28.0

# UTILITY & EVALUATION
scikit-learn>=1.3.0
tqdm>=4.65.0